fix: Remove period (`.`) from destination path
GitHub's `actions/artifact-upload` ignores dotfiles by default now (a breaking change) so this Action will not work as expected for users impacted by that change. The change here is technically breaking (as the default destination will no longer start with a `.`) but because the destination name is randomly generated, it should not impact any integrations.